requirements.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
flask-crontab==0.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PySocks==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
tweepy==3.9.0
urllib3==1.25.11
Werkzeug==1.0.1
